requirements.txt
anyio==3.5.0
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fastapi==0.74.1
h11==0.13.0
idna==3.3
pydantic==1.9.0
python-dotenv==0.19.2
requests==2.27.1
sniffio==1.2.0
soupsieve==2.3.1
starlette==0.17.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5